feat: enhance booking price distribution for program bookings and add staff role update authorization middleware

+++ routes/api_v1/staff.php
@@ -4,6 +4,7 @@ use App\Staff\Controllers\StaffController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('staff', [StaffController::class, 'index']);
-Route::post('staff/invite', [StaffController::class, 'invite']);
-Route::put('staff/{userId}/role', [StaffController::class, 'updateRole']);
-Route::delete('staff/{userId}', [StaffController::class, 'destroy']);
+Route::post('staff/invite', [StaffController::class, 'invite'])->middleware('staff.role.authorize');
+Route::put('staff/{userId}/role', [StaffController::class, 'updateRole'])->middleware('staff.role.authorize');
+Route::put('staff/{userId}/password', [StaffController::class, 'updatePassword'])->middleware('staff.role.authorize');
+Route::delete('staff/{userId}', [StaffController::class, 'destroy'])->middleware('staff.role.authorize');
